fixed structure definitions and fixed and testes moving_lift_down function
