+++ include/moving_lift_down.h
@@ -4,7 +4,8 @@
 */
 #ifndef __MOVING_LIFT_DOWN_H__
 #define __MOVING_LIFT_DOWN_H__
+#include "structures.h"
 
-extern void moving_lift_down(struct elevator elevator_arr[], unsigned int index);
+extern int moving_lift_down(struct elevator *elevator_arr, unsigned int index);
 
 #endif
