+++ src/moving_lift_down.c
@@ -5,13 +5,13 @@
 #include "moving_lift_down.h"
 
 
-void moving_lift_down(struct elevator *elevator_arr, unsigned int index){
+int moving_lift_down(struct elevator elevator_arr[], unsigned int index){
 
     if(index >= elevator_count){
-        fprintf(stderr, "index >= elevator_arr size. %d >= %d ",index,1);
-        return;
+        fprintf(stderr, "index >= elevator_arr size. %d >= %d \n",index, elevator_count);
+        return -1;
     }
     elevator_arr[index].direction_up = false;
     elevator_arr[index].direction_down = true;
-    return ;
+    return 1;
 }
